make typed hub clients public
