+++ src/SignalR/server/Core/src/TypedHubCallerClients.cs
@@ -1,0 +1,27 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using Microsoft.AspNetCore.SignalR.Internal;
+
+namespace Microsoft.AspNetCore.SignalR
+{
+    internal class TypedHubCallerClients<T> : TypedHubClients<T>, IHubCallerClients<T>
+    {
+        private readonly IHubCallerClients _hubClients;
+
+        public TypedHubClients(IHubCallerClients dynamicContext)
+            : base(dynamicContext)
+        {
+            _hubClients = dynamicContext;
+        }
+
+        public T Caller => TypedClientBuilder<T>.Build(_hubClients.Caller);
+
+        public T Others => TypedClientBuilder<T>.Build(_hubClients.Others);
+
+        public T OthersInGroup(string groupName)
+        {
+            return TypedClientBuilder<T>.Build(_hubClients.OthersInGroup(groupName));
+        }
+    }
+}
